requirements.txt
altair==5.4.1
anywidget==0.9.13
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bqplot==0.12.43
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
duckdb==1.1.1
executing==2.1.0
filelock==3.16.1
folium==0.17.0
fonttools==4.54.1
gdown==5.2.0
geojson==3.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.27.0
ipytree==0.2.2
ipyvue==1.11.1
ipyvuetify==1.10.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-leaflet==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13        
kiwisolver==1.4.7
leafmap==0.38.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.8.4
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
psygnal==0.11.1
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.8.3
python-box==7.2.0
python-dateutil==2.9.0.post0      
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scooby==0.10.0
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
whitebox==2.3.5
whiteboxgui==2.3.0
widgetsnbextension==4.0.13
xyzservices==2024.9.0